--- spin/cmds/__init__.py
@@ -1,8 +1,0 @@
-from . import meson
-
-# Backward compatibility with older versions
-build = meson.build
-test = meson.test
-ipython = meson.ipython
-python = meson.python
-shell = meson.shell
